Added Util folder and turned on Prettier to tidy file structure.
Added the Util folder to start separating UI from services and logic. There was no automated code formatting in place so Prettier was initialised. This is going to run each time it touches one of the files from the legacy app. May be useful to run as a batch over the whole project.